Move document_formatting files out of formatting/ dir
