+++ main/lsp/requests/requests.h
@@ -5,9 +5,9 @@
 #include "main/lsp/requests/code_action_resolve.h"
 #include "main/lsp/requests/completion.h"
 #include "main/lsp/requests/definition.h"
+#include "main/lsp/requests/document_formatting.h"
 #include "main/lsp/requests/document_highlight.h"
 #include "main/lsp/requests/document_symbol.h"
-#include "main/lsp/requests/formatting/document_formatting.h"
 #include "main/lsp/requests/get_counters.h"
 #include "main/lsp/requests/hover.h"
 #include "main/lsp/requests/implementation.h"
